refactor(chat-api): split conversation service into focused domains
