+++ apps/chat-api/src/conversations/tests/conversation.controller.integration.spec.ts
@@ -24,6 +24,10 @@ import {
 import { ConversationNamingService } from '../conversation-naming.service';
 import { ConversationController } from '../conversation.controller';
 import { ConversationService } from '../conversation.service';
+import { ConversationLifecycleService } from '../lifecycle/conversation-lifecycle.service';
+import { ConversationListingService } from '../listing/conversation-listing.service';
+import { ConversationPersistenceService } from '../persistence/conversation-persistence.service';
+import { ConversationStreamingService } from '../streaming/conversation-streaming.service';
 
 const TEST_USER = {
   sid: 'test-sid',
@@ -244,6 +248,10 @@ describe('ConversationController (integration)', () => {
           { provide: ConfigService, useValue: configService },
           DialClientService,
           ConversationService,
+          ConversationPersistenceService,
+          ConversationListingService,
+          ConversationLifecycleService,
+          ConversationStreamingService,
           UserConfigService,
           ScheduledTaskUnreadService,
           ConversationGenerationService,
