+++ openspec/changes/archive/2026-08-07-split-conversation-service/.openspec.yaml
@@ -1,0 +1,2 @@
+schema: spec-driven
+created: 2026-08-06
