+++ openspec/specs/backend-owned-generation-persistence/spec.md
@@ -6,7 +6,7 @@ The backend owns conversation persistence across the generation lifecycle — sa
 
 ### Requirement: Backend persists the conversation across the generation lifecycle
 
-`ConversationService.streamCompletion` (`apps/chat-api/src/conversations/conversation.service.ts`) SHALL own conversation persistence for a completion. The frontend MUST NOT call `saveConversation` during streaming. The backend SHALL save at the start of generation (user message + empty assistant placeholder), on successful completion (full assembled assistant message), and on stop/error (the partial assistant message accumulated so far).
+`ConversationStreamingService.streamCompletion` (`apps/chat-api/src/conversations/streaming/conversation-streaming.service.ts`, invoked via the `ConversationService` facade) SHALL own conversation persistence for a completion. The frontend MUST NOT call `saveConversation` during streaming. The backend SHALL save at the start of generation (user message + empty assistant placeholder), on successful completion (full assembled assistant message), and on stop/error (the partial assistant message accumulated so far).
 
 #### Scenario: Start state saved before streaming
