+++ openspec/specs/file-manager-tab-config/spec.md
@@ -159,7 +159,7 @@ The hook SHALL run a `useEffect` (dependencies: `fileManagerTabs`, `activeTab`, 
 
 ### Requirement: File-level sharedWithMe/publishedWithMe filters are explicitly waived
 
-This capability SHALL NOT implement client-side or BFF-side filtering of individual files by a `sharedWithMe`/`publishedWithMe` provenance flag. `ListFilesItemDto` and the generated `ListFilesItemDto` carry no such fields today. This is not a DIAL Core limitation — `apps/chat-api/src/conversations/conversation.service.ts` already implements the equivalent merge-and-flag pattern for conversations — but porting it to files would require introducing a combined, cross-tab listing capability that does not exist in the current per-tab-endpoint file-manager architecture (see design.md D4/Open Questions). This waiver applies only to per-item provenance flags; tab-scoped listing (each tab already fetching from its own bucket/endpoint per `file-manager-tabs`) is unaffected and continues to work as already specified.
+This capability SHALL NOT implement client-side or BFF-side filtering of individual files by a `sharedWithMe`/`publishedWithMe` provenance flag. `ListFilesItemDto` and the generated `ListFilesItemDto` carry no such fields today. This is not a DIAL Core limitation — `apps/chat-api/src/conversations/listing/conversation-listing.service.ts` already implements the equivalent merge-and-flag pattern for conversations — but porting it to files would require introducing a combined, cross-tab listing capability that does not exist in the current per-tab-endpoint file-manager architecture (see design.md D4/Open Questions). This waiver applies only to per-item provenance flags; tab-scoped listing (each tab already fetching from its own bucket/endpoint per `file-manager-tabs`) is unaffected and continues to work as already specified.
 
 #### Scenario: No per-file provenance field is introduced
